Enable fast forwarding (fast routing).
A description of fastforwarding:

Fastforwarding caches the results of a route lookup for destination
addresses that are not on the local machine, and uses the cached route
to short-circuit the normal (relatively slow) route lookup process.  The
packet flows directly from one layer2 input routine directly to the opposing
layer2 output routine without traversing the IP layer.