Yet-another-submission-by: databeestje (Seth) (who will have a cvs bit soon)
CVSTrac message:
This should fix it and might even work with https reading from the XML_RPC_Client code.

Tested with http on port 8888 on both carp boxen.

We need to state this in the Wiki Somewhere.

For syncing configuration both carp machines need to be running the same protocol and on the same port to function.

Teehee!