Remove the parent interface reconfiguration functions they are only bogusly/erroneously called here: * On bootup without this change the parent interface will be configured twice. * Even when the system is up a vlan does not have to mock up with the parent apart guaranteeing that it is up.
The only thing that does not work is makeing a VLAN on dynamic interfaces like pppoe/pptp.
Afaik it has never worked and i do not think that many people would use vlans over pppoe,
if a feature request for such things comes up it will be on my TODO around 2.x of pfSense.