this is of no use with if_bridge, in the sense it was with m0n0 (where this came from)
Since we force filtering on the parent interfaces, we cannot disable it in the same sense as in m0n0 for the bridge.

Because you cannot assign a bridge as an interface, you can't filter on it anyway. Will now always remain disabled and checkbox on Advanced page will be ignored.