Added custom_php_service_status_command support for packages.
Package authors can define a
  custom_php_service_status_command
block in their package XML. Such block should only contain function
calls, and the first should be one to include their function library
file if necessary, as this is not done by service-utils.inc.

The return code of the last function call will be returned to the service
check.