Do not add hostname to watch in the refresh ipsec policy section, there may be other tunnels using the same endpoint which need refreshing as well. This is also done in the part where the racoon configuration is written so it's safe to skip it here.
Silence the logging in the dnscache code and the ipsec route add code into a debug check

FIXME: there is a possible situation where hostname changes can be missed. In a lot of tunnels situation there
might be updates of other hostnames that changed while we were reconfiguring. Not sure how to handle that.