Added checks to wol command execution to make it more obvious when the script fails.
If the wol command returns with anything other than 0, alert the user that the command
failed, and that they should look at the system log to find out why the command failed.
I'm adding this out of frustration with the fact that the wol command hasn't been included
at all in recent snapshots, and because the webgui doesn't give any indication of the command
not working, it took much longer for me to figure out why it wasn't working.  I hope
this can save someone else time.