I have PPP with mpd5 working really well now, including setting the APN and APN Number, and the SIM PIN and wait time for the SIM to settle after the PIN is entered.
Still no work on multilink PPP with serial port links.

Right now the code creates the mpd.secret file. I tested it using the set auth
password <password> syntax in the mpd.conf file and got errors so I kept the
mpd.secret file generation. PPP for modems does need the mpd.script file. It's
the chat script file for the link establishment. Right now it has to be in
/usr/local/sbin/mpd.script, and it's linked to /var/etc/ at boot time when ppp
is first configured. I also took away the "Dialcmd" field in the PPP setup page
(and a couple others) because the script is rather comprehensive and is designed
to handle many modems. My two 3G modems worked without modification to the
script. I did have to modify the scrip to handle the setting of APN and SIM
PINs.

I made the Init String field work, and the default init commands in the ATT&T,
Verizion, and Sprint defaults now come up in the Init String field. You also no
longer should put an "AT" at the beginning of modem commands entered in the web
page because the script does that for you. I implemented interface renaming and
that simplified the new code a lot, so that's nice. Now ppp interfaces start
with ppp0 and go up as you add more.

This patch refers to the ppp-linkup and linkdown scripts in the /usr/local/sbin/
directory so custom PPP linkup and linkdown scripts are not needed anymore.

The persistent uptime tracking is still an open issue, but it works within each
connected session. I'm already experiencing much smother conditions with mpd5
for PPP on 3G than I was with userland ppp. It's very nice.