Don't use "local" as a domain. It breaks DNS resolution for hosts running mDNS.
The "local" search domain signifies to local hosts that are running
mDNS (bonjour or avahi) that mDNS is to be used to look up local hosts
instead of doing a normal DNS query to the server listed in
/etc/resolv.conf on the local host.

Also, hosts running bonjour or avahi can not ping or reach *any* other
host by name if it is not running some mDNS implementation.
Essentially, if DHCP tells the local host that it's search domain is
"local" then hosts running mDNS will not query the pfsense DNS server
for any local lan DNS lookups.

See here on apple's website:
http://support.apple.com/kb/TA20999?viewlocale=en_US

Quote from above link:
"To indicate that the name should be looked up using local multicast
instead of a standard DNS query, all Bonjour host names end with the
extension ".local."