Fix all issues that Ermal found.
Interfaces.php problems.
1- When creating a pppoe ptpid on the config is empty(<<ptpid/>).
2- after switching from pppoe to pptp in the config there remains
pppoe-reset-type which is not that much of a bug but needs to be
fixed.
3- When choosing ppp type even though i do not have a device i am
allowed to save. After saving it switches to none type automatically.
The <ppp> section still has values of pptp config and some of pppoe
and only the apn related to ppp. The interface assignment still has
teh interface pptp1! (This is quite buggy i would say.)

Interfaces -> assign
1- After doing all the tests on interfaces.php everything is shown
with ppp() on interfaces.php picture attached. (probably an issue with
how you do find ppp interfaces around.)
2- MPD is started even though the interface is not enabled after assignment.
3- During assignment as i tell you even below remove teh pptpid from
the interface an duse interface name.

Interfaces -> assign -> ppp
1- I have an emtpy thing in there when i do all the actions on interfaces.php