Update wizard files for setup wizard to write proper config for current ppp architecture.
I also allowed selection of physical interface from the wizard screen. This is crucial
because if the user has the WAN set to anything other than the same phyical interface used
by pppoe/pptp, the old wizard has no way of knowing what physical port to use, and the wizard
can generate a bad config. Feel free to expand the list of available interfaces in the dropdown
for interface selection.