Remove uncessesary function call.
This function call is extraneous. mpd will automatically shut down existing mpd processes
gracefully because we call it with the "-k" command line switch.

In addition, this call causes mpd to be started twice in rapid succession when users have
"dial on demand" enabled. This can lead to problems with orphaned netgraph nodes which can
only be cleaned up via reboot.