Initial support for Cisco IPSec over GRE tunnels
The current state of the ipsec transport mode w/ gre tunnels is NOT compatible with Cisco routers / VPN concentrators.

Cisco expects :

1) only GRE packets should be encrypted  not ANY
2) Only outgoing packets coming from local -> remote of type GRE should match on each end (not bidirectional).

The updated php and inc files take care of making a compatible tunnel, but it appears to break the IPSec monitoring on PFSense (the tunnel DOES work though).

Feel free to email me if you have any questions, this was tested with IOS 12.4.x on both 7200 and 2921 routers.

Signed-off-by: John Grange <john@sd-networks.net>