Tidy up percentage logging to console
When on the console, the code only intended to update the percentage downloaded every 10%. Due to string-and-int type differences, the test was not working, all percentages were being logged. This is fixed.
Also, add a newline after 100% - then whatever outputs next gets a clean start on a new line of output.
Tidies up the console output during a post-firmware-upgrade package install.