Fix deleting DHCP leases
This broke when the code was changed to suck the whole leases file into an array with:
$leases_contents = file($leasesfile);
Each array entry has the text of a line in the file AND the newline at the end of the line. So when matching array entries, the match has to expect the "\n" on the end.