Do not translate keywords like "default"
Keywords that are in drop-down lists that get put straight into the config files and passed into conf files should not be translated. They are not really "English", they are keywords for the pieces of software that use them.
These ones were causing problems when trying to setup Captive Portal in Portuguese.