Handle null parameter to services_dyndns_configure
When service_dyndns_configure is called with a null parameter, it should check/update the dyndns state of all interfaces configured for dyndns. But actually it is only updating gateway groups. e.g. the daily cron job /etc/rc.dyndns.update does not update an ordinary WAN interface.
This fixes the problem.