Update usr/local/www/services_dhcp.php
Inform user how the Primary/Secondary DHCPd process is determined in a failover pair so they don't end up with two secondary servers. For example, when using advskew=64 (and advskew=164 if using settings sync to peer for carp ips) on the carp master machine which they *think* will be the DHCP primary.