Interface assignment wrongly changing:
When WAN is set to PPPoE and user set other interfaces IP address using
console, it wrongly change the interface assignment to use the same
device of wan. It was caused by a hard coded "wan" on
console_get_interface_from_ppp() call, when it should use $interface
instead. It should fix #2074