Update all the "include" files .
Update all the "include" files so that the HTML generated is W3C
compliant.

This makes it easier to update all the PHP and INC files later to W3C
standard.

&lt;img&gt; element not empty or not closed
&lt;img&gt; lacks "alt"
attribute
&lt;table&gt; lacks "summary" attribute
&lt;br&gt; element not
empty or not closed
&lt;input&gt; element not empty or not
closed
&lt;a&gt; attribute "target" has invalid value "_new"

Updated the "loader.js" and "bottom-loader.js" in themes directory due
to the update in "/usr/local/www/guiconfig.inc"