Make synamic firewall logs and widget add new rows:
The Dynamic Firewall Log page doesn't add rows, it only updates the php
generated rows which means that if it's loaded after the log is cleared,
it'll never show any activity. The same occurs for the log widget.

This pacth fix this issue moving from absolutely sized spans to tables.

It should fix #2763