Validate interface IP explicitly as v4 or v6
The IP Address fields on the Interfaces Edit page were just using is_ipaddr to validate them. I could put an IPv4 address in as the IPv6 static IP and it took it and saved it, then got my system in a knot! It even adjusted the subnet bits dropdown when I typed a valid IPv4 address format in the IPv6 filed.
And IPv6 addresses could be put in IPv4 fields also.