Cleanup code handling 25 day dyndns update
The code that checks if an IP has not changed for 25 days has out-of-date comments and uses "25" in multiple places and the magic number for 25 days in seconds. Clean this up so that "25" is set once in a variable and used everywhere else for logging and calculating seconds. This makes the code easier to read and maintain, particularly if in future the max age needs to be changed for some/all providers.