Add option to display host name
If hostname display is requested, then get locally-known IP to hostname mappings from dnsmasq and dhcp static configs. Replace IP addresses with the hostname (if known).
Note: no attempt is made to implement reverse lookups etc as that would take far too much time to get the answers or timeout. This feature is intended as a quick and easy way to get names onto the bandwidth by ip table for sites that do DHCP from pfSense and allocate static mappings.