Services status widget display 1st sentence of description
Some packages (e.g. bandwidthd, dansguardian, HAVP, nmap, squid3...) have a very verbose package description. This is displayed in the description column of the services status widget. That makes it take a lot of space on the dashboard. IMHO the first sentence gives a good concise description, so this change just displays the first sentence (text before a "."). Note: I purposely left the "." out of the string, I think it looks neater along with the descriptions of other built-in services - "DHCP Service" "DNS Forwarder"...