DNS Forwarder allow null forwarding address
dnsmasq allows a blank forwarding IP address to be specified in --server parameters. In that case, dnsmasq will attempt a local lookup of the name (e.g. in hosts file), but will not forward the request on to the standard name servers.
This is a useful way to prevent internal name requests from being sent out to the real internet name servers.