Fixup gwlb functions related to gw failover
Make apinger pass the gateway parameter both when an alarm alarms (gateway is down) and when it comes good (gateway is up). This way the downstream openvpn and dyndns code can know what happened and do sensible stuff.
Return the friendly interface name ("wan", "opt1" etc) rather than the actual device name ("vr0", "vr1" etc). The openvpn and dynamic dns code that uses this needs to work with various entries in the config that all use the friendly interface name.