Some changes in functions involved with radius!
On line 1204 "if (is_null($pipeno)) {" replaced for "if (empty($pipeno)) {". Pineno takes value 0 if there are no free pipes, is_null would return false instead of true in case it is 0

I guess there is an error in the function portal_mac_radius specifically in the following condition:

if (! empty ($ auth_list ['url_redirection']))
portal_reply_page ($ auth_list ['url_redirection'], "redir");
     return FALSE;

This condition appears at the end of the function so the function never returns TRUE but FALSE always, based on the following:

When using if statements without the curly braces, remember than only one statement will be Executed as part of That condition. If you want to place multiple statements you must use curly braces, and not just put them on the same line.

if (1 == 0) echo "Test 1." echo "Test 2";


If i'm wrong discar the changes.

Thank you!