Make NAT port forward data entry error message match GUI text
The GUI has "Redirect target IP" and "Redirect target port". But the error message when validating "Redirect target port" refers to it as "local port". I was just confused when I forgot to type in the "Redirect target port", the error told me I had to enter a valid local port. I looked down the GUI page for "local port" and it is not there.