Display pfSense interface name on status interfaces
Sometimes it is a bit difficult to work out which interface descriptive name is which pfSense internal interface name (wan, lan, opt1, opt2...). It is useful to know for:
a) hardware that has wan, lan, opt1... pre-printed on the case (some Alix cases for example), making it easier for less-trained remote office staff to work out which cable corresponds to which interface in pfSense.
b) looking in config.xml or other techo places where the internal pfSense name is stored/referenced - knowing that straight from a status GUI display saves having to first find the descriptive name in the config, then the pfSense internal name.
What do you think? Is this a worthwhile place to display this data?