Allow setting a default scale type preference for the traffic graphs widget
I originally submitted this at https://redmine.pfsense.org/issues/2994
but it seems that is not used for commits now so I am adding it here.
It works great on my box, and I can't see these changes causing any
issues on other types.

I like to have Autoscale (follow) set for the traffic graphs widget and
hate having to click on Autoscale (up) to change it to follow every
time I enter the dashboard page.

This preference is written to the traffic graph widget section in
config.xml