Adjust DHCP lease local time back to UTC if required
The time in dhcpd.leases file is now actually in local time. So if the user has selected to display the leases in local time, then nothing needs to be done. And if they want to display in UTC, then the time zone offset needs to be subtracted from the time.
Forum post: http://forum.pfsense.org/index.php/topic,70504.0.html