Gateway Monitor Advanced Settings
Exposes 3 additional apinger configuration options in the gateway monitor advanced section which can either be set manually, calculated based on interval, or use the hard coded defaults.
	1) Avg Delay Samples
	2) Avg Loss Samples
	3) Avg Loss Delay Samples

The current hard coded defaults are okay for a short interval such as the 1 second default.  But do not work well when a longer interval is required.